Added a section describing filesystem structure
I added a section to the README which outlines the default filesystem structure used by LEClient, as well as the various customisation options available and how they interact with each other. I also removed a (presumably outdated?) reference to `getCertificate()` storing certificates under a particular basename, which does not seem to be the case as of the current software version.